fix(panel): Fixed panel issues and appbar,navbar buttons
Closes #301
