+++ packages/geoview-core/src/core/components/map/map.tsx
@@ -15,8 +15,6 @@ import { Crosshair } from "../crosshair/crosshair";
 import { MousePosition } from "../mouse-position/mouse-position";
 import { Attribution } from "../attribution/attribution";
 import { Snackbar } from "../../../ui/snackbar/snackbar";
-import { Appbar } from "../appbar/app-bar";
-import { NavBar } from "../navbar/nav-bar";
 import { NorthArrow, NorthPoleFlag } from "../north-arrow/north-arrow";
 import { ClickMarker } from "../click-marker/click-marker";
 
@@ -150,9 +148,10 @@ export function Map(props: TypeMapConfigProps): JSX.Element {
     );
 
     return () => {
+      api.event.off(EVENT_NAMES.EVENT_MAP_ADD_COMPONENT, id);
+      api.event.off(EVENT_NAMES.EVENT_MAP_REMOVE_COMPONENT, id);
       api.event.off(EVENT_NAMES.EVENT_BASEMAP_LAYERS_UPDATE, id);
     };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   return (
@@ -230,20 +229,10 @@ export function Map(props: TypeMapConfigProps): JSX.Element {
           {deviceSizeMedUp && <MousePosition id={id} />}
           <ScaleControl position="bottomright" imperial={false} />
           {deviceSizeMedUp && <Attribution attribution={attribution} />}
-          <div
-            className="leaflet-control cgp-appbar"
-            style={{
-              boxSizing: "content-box",
-              zIndex: defaultTheme.zIndex.appBar,
-            }}
-          >
-            <Appbar />
-          </div>
           <NorthArrow projection={crs!} />
           <NorthPoleFlag projection={crs!} />
           <Crosshair id={id} />
           <ClickMarker />
-          <NavBar />
           <SnackbarProvider
             maxSnack={3}
             dense
