+++ packages/geoview-core/src/geo/map/map.ts
@@ -112,6 +112,7 @@ export class MapViewer {
     this.currentPosition = new LatLng(mapProps.center[0], mapProps.center[1]);
 
     this.appBarButtons = new AppbarButtons(this.id);
+    this.navBarButtons = new NavbarButtons(this.id);
   }
 
   /**
@@ -241,9 +242,7 @@ export class MapViewer {
   /**
    * Function called when the map has been rendered and ready to be customized
    */
-  mapReady = (): void => {
-    this.navBarButtons = new NavbarButtons(this.id);
-  };
+  mapReady = (): void => {};
 
   /**
    * Change the language of the map
