+++ packages/geoview-core/src/ui/panel/panel-api.ts
@@ -4,7 +4,6 @@ import { api } from "../../api/api";
 import { EVENT_NAMES } from "../../api/event";
 import { CheckboxListAPI } from "../list/checkbox-list/checkbox-list-api";
 import { TypePanelProps } from "../../core/types/cgpv-types";
-import { DefaultPanel } from ".";
 
 /**
  * Class used to handle creating a new panel
@@ -14,22 +13,22 @@ import { DefaultPanel } from ".";
  */
 export class PanelApi {
   // panel type (appbar, navbar)
-  type: "appbar" | "navbar";
+  type: string | undefined;
 
   // panel open status (open/closed)
-  status?: boolean;
+  status: boolean | undefined;
 
   // width of the panel
-  width?: string | number;
+  width: string | number;
 
   // panel header icon
-  icon?: React.ReactNode | Element;
+  icon: React.ReactNode | Element;
 
   // panel header title
-  title?: string;
+  title: string;
 
   // panel body content
-  content?: React.ReactNode | Element;
+  content: React.ReactNode | Element;
 
   // the linked button id that will open/close the panel
   buttonId: string;
@@ -44,13 +43,13 @@ export class PanelApi {
    *
    * @param {TypePanelProps} panel the passed in panel properties when panel is created
    * @param {string} buttonId the button id of the button that will manage the panel
-   * @param {string} mapId the id of the map that this panel belongs to
+   * @param {string} mapId the map id that this panel belongs to
    */
   constructor(panel: TypePanelProps, buttonId: string, mapId: string) {
     this.mapId = mapId;
     this.buttonId = buttonId;
     this.type = panel.type;
-    this.title = panel.title;
+    this.title = panel.title || "";
     this.icon = panel.icon;
     this.content =
       panel.content !== undefined && panel.content !== null
@@ -60,7 +59,7 @@ export class PanelApi {
       panel.status !== undefined && panel.status !== null
         ? panel.status
         : false;
-    this.width = panel.width;
+    this.width = panel.width || 300;
   }
 
   /**
@@ -69,18 +68,36 @@ export class PanelApi {
   open = (): void => {
     this.status = true;
 
-    this.updatePanel({
-      title: this.title,
-      width: this.width,
-      content: this.content,
-      status: this.status,
-      icon: this.icon,
-      type: this.type,
-    });
+    // close all other panels
+    this.closeAll();
 
     api.event.emit(EVENT_NAMES.EVENT_PANEL_OPEN, this.mapId, {
+      handlerId: this.mapId,
       buttonId: this.buttonId,
     });
+  };
+
+  /**
+   * Close all other panels
+   */
+  closeAll = (): void => {
+    Object.keys(api.map(this.mapId).appBarButtons.buttons).map(
+      (groupName: string) => {
+        // get button panels from group
+        const buttonPanels = api.map(this.mapId).appBarButtons.buttons[
+          groupName
+        ];
+
+        // get all button panels in each group
+        Object.keys(buttonPanels).map((buttonId) => {
+          const buttonPanel = buttonPanels[buttonId];
+
+          if (this.buttonId !== buttonPanel.id) {
+            buttonPanel.panel?.close();
+          }
+        });
+      }
+    );
   };
 
   /**
@@ -89,9 +106,8 @@ export class PanelApi {
   close = (): void => {
     this.status = false;
 
-    this.updatePanel(DefaultPanel.panel, false);
-
     api.event.emit(EVENT_NAMES.EVENT_PANEL_CLOSE, this.mapId, {
+      handlerId: this.mapId,
       buttonId: this.buttonId,
     });
   };
@@ -150,48 +166,16 @@ export class PanelApi {
    * Change the content of the panel
    *
    * @param {React Element} content the content to update to
-   * @param {boolean} override an optional boolean value that will use the passed in content as default panel content (default true)
    *
    * @returns {Panel} this panel
    */
-  changeContent = (
-    content: React.ReactNode | Element,
-    override: boolean = true
-  ): PanelApi => {
-    if (override) this.content = content;
+  changeContent = (content: React.ReactNode | Element): PanelApi => {
+    this.content = content;
 
     api.event.emit(EVENT_NAMES.EVENT_PANEL_CHANGE_CONTENT, this.mapId, {
+      handlerId: this.mapId,
+      buttonId: this.buttonId,
       content,
-      buttonId: this.buttonId,
-    });
-
-    return this;
-  };
-
-  /**
-   * Update the panel with new values
-   *
-   * @param {TypePanelProps} panelProps the updated panel properties
-   * @param {boolean} override an optional boolean value that will use the passed in props as default panel props (default true)
-   *
-   * @returns {Panel} this panel
-   */
-  updatePanel = (
-    panelProps: TypePanelProps,
-    override: boolean = true
-  ): PanelApi => {
-    if (override) {
-      this.status = panelProps.status || this.status;
-      this.title = panelProps.title || this.title;
-      this.icon = panelProps.icon || this.icon;
-      this.type = panelProps.type || this.type;
-      this.width = panelProps.width || this.width;
-      this.content = panelProps.content || this.content;
-    }
-
-    api.event.emit(EVENT_NAMES.EVENT_PANEL_UPDATE, this.mapId, {
-      panelProps,
-      buttonId: this.buttonId,
     });
 
     return this;
